+++ test/integration/test_autocomplete.py
@@ -15,7 +15,7 @@ import pytest
 
 from test.integration.helpers import skip_on_windows
 
-TIMEOUT = 10
+TIMEOUT = 20  # CI can be slow at times
 
 BASHRC_CONTENT = """\
 # ~/.bashrc dummy file